requirements.txt
torch==2.1.2
torchvision==0.16.2
diffusers==0.26.3
transformers==4.39.3
accelerate==0.27.2
peft==0.9.0
datasets==2.19.1
huggingface_hub==0.22.2
Pillow==10.2.0